el_code_location/requirements.txt
dagster-postgres
dagster-embedded-elt
dagster-aws
dlt[motherduck]